# Planner/Planner/api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
...
  /api/eolopark/{id}:
    get:
      tags:
      - api-controller
      operationId: getEolopark
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EoloPark'
    put:
      tags:
      - api-controller
      operationId: updateEolopark
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EoloPark'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EoloPark'
    delete:
      tags:
      - api-controller
      operationId: deleteEolopark
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EoloPark'
  /api/aerogenerator/{id}:
    get:
      tags:
      - api-controller
      operationId: getaerogenerator
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Aerogenerator'
    put:
      tags:
      - api-controller
      operationId: updateAerogenerator
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Aerogenerator'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Aerogenerator'
    delete:
      tags:
      - api-controller
      operationId: deleteaerogenerator
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Aerogenerator'
...
  /api/eolopark:
    get:
      tags:
      - api-controller
      operationId: getEoloparks
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageEoloPark'
    post:
      tags:
      - api-controller
      operationId: createEolopark
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EoloPark'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EoloPark'
...
components:
  schemas:
    Aerogenerator:
      type: object
      properties:
        eoloPark:
          $ref: '#/components/schemas/EoloPark'
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        bladeLength:
          type: number
          format: double
        height:
          type: number
          format: double
        power:
          type: number
          format: double
        id:
          type: string
        trueId:
          type: integer
          format: int64
    EoloPark:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        city:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        area:
          type: number
          format: double
        terrainType:
          type: string
          enum:
          - PLAIN
          - HILL
          - MOUNTAIN
          - DESERT
          - FOREST
        owner:
          type: integer
          format: int64
        aerogenerator:
          $ref: '#/components/schemas/Aerogenerator'
...
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/SortObject'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...
    PageEoloPark:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/EoloPark'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        empty:
          type: boolean
...